requirements.txt
black==21.11b0
Pillow==8.4.0